# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AffectLog Trustworthy AI Assessment Building Block (ALT-AI)
  description: API for dataset profiling, xAPI normalization, fairness metrics, model
    explainability, and EU AI Act / GDPR compliance exports. Aligned with D3.7 EDGE-Skills
    / Prometheus-X.
  contact:
    name: AffectLog
    url: https://github.com/roy-saurabh/edge_affectlog
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 0.1.0
paths:
...
  /v1/datasets/validate:
    post:
      tags:
      - Datasets
      summary: Validate CSV schema
      operationId: validate_dataset_v1_datasets_validate_post
      parameters:
      - name: x-request-id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Request-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasetValidateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetValidateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/datasets/ingest:
    post:
      tags:
      - Datasets
      summary: Register a dataset for processing
      operationId: ingest_dataset_v1_datasets_ingest_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasetIngestRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetIngestResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/datasets/{dataset_id}:
    get:
      tags:
      - Datasets
      summary: Get dataset info
      operationId: get_dataset_v1_datasets__dataset_id__get
      parameters:
      - name: dataset_id
        in: path
        required: true
        schema:
          type: string
          title: Dataset Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Dataset V1 Datasets  Dataset Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/datasets/{dataset_id}/transform:
    post:
      tags:
      - Datasets
      summary: Transform Dataset
      operationId: transform_dataset_v1_datasets__dataset_id__transform_post
      parameters:
      - name: dataset_id
        in: path
        required: true
        schema:
          type: string
          title: Dataset Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasetTransformRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetTransformResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/datasets/{dataset_id}/profile:
    post:
      tags:
      - Datasets
      summary: Profile a dataset
      operationId: profile_dataset_v1_datasets__dataset_id__profile_post
      parameters:
      - name: dataset_id
        in: path
        required: true
        schema:
          type: string
          title: Dataset Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Profile Dataset V1 Datasets  Dataset Id  Profile Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/audits/run:
    post:
      tags:
      - Audits
      summary: Start an audit pipeline run
      operationId: run_audit_endpoint_v1_audits_run_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuditRunRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditRunResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/audits/{run_id}:
    get:
      tags:
      - Audits
      summary: Get audit run status
      operationId: get_audit_v1_audits__run_id__get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditRunResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/audits/{run_id}/artifacts:
    get:
      tags:
      - Audits
      summary: List run artifacts
      operationId: get_artifacts_v1_audits__run_id__artifacts_get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Artifacts V1 Audits  Run Id  Artifacts Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/audits/{run_id}/sop:
    get:
      tags:
      - Audits
      summary: Get SOP markdown
      operationId: get_sop_v1_audits__run_id__sop_get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            text/plain:
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/audits/{run_id}/compliance-graph:
    get:
      tags:
      - Audits
      summary: Get compliance JSON-LD graph
      operationId: get_compliance_graph_v1_audits__run_id__compliance_graph_get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Compliance Graph V1 Audits  Run Id  Compliance
                  Graph Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/compliance/jsonld:
    post:
      tags:
      - Compliance
      summary: Build and return compliance JSON-LD graph
      operationId: build_jsonld_v1_compliance_jsonld_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplianceJSONLDRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Build Jsonld V1 Compliance Jsonld Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/compliance/data-card:
    post:
      tags:
      - Compliance
      summary: Build Data Card for a run
      operationId: build_data_card_endpoint_v1_compliance_data_card_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DataCardRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Build Data Card Endpoint V1 Compliance Data Card Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/compliance/model-card:
    post:
      tags:
      - Compliance
      summary: Build Model Card for a registered model
      operationId: build_model_card_endpoint_v1_compliance_model_card_post
      parameters:
      - name: model_id
        in: query
        required: true
        schema:
          type: string
          title: Model Id
      - name: run_id
        in: query
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Build Model Card Endpoint V1 Compliance Model Card
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AuditRunRequest:
      properties:
        input_path:
          type: string
          title: Input Path
        recipe:
          type: string
          title: Recipe
          default: configs/recipes/maskott_tactileo.yaml
        output_dir:
          anyOf:
          - type: string
          - type: 'null'
          title: Output Dir
        chunk_size:
          type: integer
          title: Chunk Size
          default: 100000
      type: object
      required:
      - input_path
      title: AuditRunRequest
      example:
        input_path: data/processed/maskott_1m.normalized.jsonl
        recipe: configs/recipes/maskott_tactileo.yaml
    AuditRunResponse:
      properties:
        run_id:
          type: string
          title: Run Id
        status:
          type: string
          title: Status
        recipe:
          type: string
          title: Recipe
        artifacts:
          additionalProperties:
            type: string
          type: object
          title: Artifacts
          default: {}
      type: object
      required:
      - run_id
      - status
      - recipe
      title: AuditRunResponse
...
    ComplianceJSONLDRequest:
      properties:
        run_id:
          type: string
          title: Run Id
        dataset_name:
          type: string
          title: Dataset Name
          default: unknown
      type: object
      required:
      - run_id
      title: ComplianceJSONLDRequest
    DataCardRequest:
      properties:
        run_id:
          type: string
          title: Run Id
        dataset_name:
          type: string
          title: Dataset Name
      type: object
      required:
      - run_id
      - dataset_name
      title: DataCardRequest
    DatasetIngestRequest:
      properties:
        file_path:
          type: string
          title: File Path
        schema_name:
          type: string
          title: Schema Name
          default: maskott_csv_v1
        recipe:
          type: string
          title: Recipe
          default: maskott_tactileo
      type: object
      required:
      - file_path
      title: DatasetIngestRequest
    DatasetIngestResponse:
      properties:
        dataset_id:
          type: string
          title: Dataset Id
        file_path:
          type: string
          title: File Path
        status:
          type: string
          title: Status
      type: object
      required:
      - dataset_id
      - file_path
      - status
      title: DatasetIngestResponse
    DatasetTransformRequest:
      properties:
        recipe:
          type: string
          title: Recipe
          default: configs/recipes/maskott_tactileo.yaml
        output_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Output Path
        chunk_size:
          type: integer
          title: Chunk Size
          default: 100000
        template_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Template Path
      type: object
      title: DatasetTransformRequest
    DatasetTransformResponse:
      properties:
        dataset_id:
          type: string
          title: Dataset Id
        output_path:
          type: string
          title: Output Path
        total_rows_in:
          type: integer
          title: Total Rows In
        total_rows_out:
          type: integer
          title: Total Rows Out
        error_count:
          type: integer
          title: Error Count
      type: object
      required:
      - dataset_id
      - output_path
      - total_rows_in
      - total_rows_out
      - error_count
      title: DatasetTransformResponse
    DatasetValidateRequest:
      properties:
        file_path:
          type: string
          title: File Path
          description: Absolute or relative path to the input file
        schema_name:
          type: string
          title: Schema Name
          description: Named schema to validate against
          default: maskott_csv_v1
      type: object
      required:
      - file_path
      title: DatasetValidateRequest
      example:
        file_path: data/samples/maskott_csv_sample.csv
        schema_name: maskott_csv_v1
    DatasetValidateResponse:
      properties:
        valid:
          type: boolean
          title: Valid
        schema_name:
          type: string
          title: Schema Name
        actual_columns:
          items:
            type: string
          type: array
          title: Actual Columns
          default: []
        missing_columns:
          items:
            type: string
          type: array
          title: Missing Columns
          default: []
        extra_columns:
          items:
            type: string
          type: array
          title: Extra Columns
          default: []
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      type: object
      required:
      - valid
      - schema_name
      title: DatasetValidateResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError